REFACTOR separate random_lists library
