+++ doc/FabLabKasse.libs.rst
@@ -13,6 +13,18 @@ Subpackages
     FabLabKasse.libs.process_helper
     FabLabKasse.libs.pxss
 
+Submodules
+----------
+
+FabLabKasse.libs.random_lists module
+------------------------------------
+
+.. automodule:: FabLabKasse.libs.random_lists
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
 Module contents
 ---------------
